+++ starkbank/invoice/__init__.py
@@ -2,3 +2,4 @@ from .__invoice import create, get, query, page, update, qrcode, pdf, payment
 from .log.__log import Log
 from . import log
 from .__payment import Payment
+from .rule.__rule import Rule